Ajustado codigo com base na arquitetura. Ajustado para que os fragments e o refresh funcionem juntos. Ajustado apresentação de layout. Processo feito apenas para o fragment 1year.
