+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:theme="@style/Base.Theme.CoinWise"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.activity.MainActivity"
+            android:name=".presentation.MainActivity"
             android:exported="false" />
         <activity
             android:name=".presentation.SplashScreen"
