+++ app/src/main/java/com/app/coinwise/presentation/SplashScreen.kt
@@ -6,7 +6,6 @@ import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
 import com.app.coinwise.R
-import com.app.coinwise.presentation.activity.MainActivity
 
 class SplashScreen : AppCompatActivity() {
