+++ app/src/main/res/layout/activity_main.xml
@@ -4,117 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.activity.MainActivity">
+    tools:context=".presentation.MainActivity">
 
-    <Button
-        android:id="@+id/button_7days_fragment_test"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="7D"
-        android:textColor="@color/black"
-        app:layout_constraintBottom_toTopOf="@id/fragment_container_view"
-        app:layout_constraintEnd_toStartOf="@id/button_30days_fragment_test"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
-        android:id="@+id/swipeToRefresh"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-            <TextView
-                android:id="@+id/tv_percentage"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="60dp"
-                android:layout_marginEnd="18dp"
-                android:textSize="20sp"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                android:text="14%" />
-
-            <ImageView
-                android:id="@+id/img_down"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="62dp"
-                android:layout_marginEnd="2dp"
-                android:src="@drawable/baseline_keyboard_double_arrow_down_24"
-                android:visibility="gone"
-                app:layout_constraintEnd_toStartOf="@+id/tv_percentage"
-                app:layout_constraintTop_toTopOf="parent" />
-
-            <ImageView
-                android:id="@+id/img_up"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="62dp"
-                android:layout_marginEnd="2dp"
-                android:src="@drawable/baseline_keyboard_double_arrow_up_24"
-                android:visibility="gone"
-                app:layout_constraintEnd_toStartOf="@+id/tv_percentage"
-                app:layout_constraintTop_toTopOf="parent" />
-
-            <com.github.mikephil.charting.charts.LineChart
-                android:id="@+id/line_chart_bitcoin"
-                android:layout_width="match_parent"
-                android:layout_height="300dp"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent" />
-    <Button
-        android:id="@+id/button_30days_fragment_test"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="30D"
-        android:textColor="@color/black"
-        app:layout_constraintBottom_toTopOf="@id/fragment_container_view"
-        app:layout_constraintEnd_toStartOf="@id/button_60days_fragment_test"
-        app:layout_constraintStart_toEndOf="@id/button_7days_fragment_test"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <Button
-        android:id="@+id/button_60days_fragment_test"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="60D"
-        android:textColor="@color/black"
-        app:layout_constraintBottom_toTopOf="@id/fragment_container_view"
-        app:layout_constraintEnd_toStartOf="@id/button_90days_fragment_test"
-        app:layout_constraintStart_toEndOf="@id/button_30days_fragment_test"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <Button
-        android:id="@+id/button_90days_fragment_test"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="90D"
-        android:textColor="@color/black"
-        app:layout_constraintBottom_toTopOf="@id/fragment_container_view"
-        app:layout_constraintEnd_toStartOf="@id/button_1year_fragment_test"
-        app:layout_constraintStart_toEndOf="@id/button_60days_fragment_test"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <Button
-        android:id="@+id/button_1year_fragment_test"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="1Y"
-        android:textColor="@color/black"
-        app:layout_constraintBottom_toTopOf="@id/fragment_container_view"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@id/button_90days_fragment_test"
-        app:layout_constraintTop_toTopOf="parent" />
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragment_container_view"
@@ -164,10 +55,10 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-        </androidx.constraintlayout.widget.ConstraintLayout>
 
 
-    </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
+
+
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
