+++ app/src/main/res/layout/fragment_graph1_year.xml
@@ -6,6 +6,55 @@
     android:layout_height="match_parent"
     tools:context=".presentation.fragment.Graph1YearFragment">
 
+    <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
+        android:id="@+id/swipeToRefresh"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent">
+
+
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
+
+        <TextView
+            android:id="@+id/tv_percentage"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="60dp"
+            android:layout_marginEnd="18dp"
+            android:textSize="20sp"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            android:text="14%" />
+
+        <ImageView
+            android:id="@+id/img_down"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="62dp"
+            android:layout_marginEnd="2dp"
+            android:src="@drawable/baseline_keyboard_double_arrow_down_24"
+            android:visibility="gone"
+            app:layout_constraintEnd_toStartOf="@+id/tv_percentage"
+            app:layout_constraintTop_toTopOf="parent" />
+
+        <ImageView
+            android:id="@+id/img_up"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="62dp"
+            android:layout_marginEnd="2dp"
+            android:src="@drawable/baseline_keyboard_double_arrow_up_24"
+            android:visibility="gone"
+            app:layout_constraintEnd_toStartOf="@+id/tv_percentage"
+            app:layout_constraintTop_toTopOf="parent" />
+
+
+
     <com.github.mikephil.charting.charts.LineChart
         android:id="@+id/line_chart_bitcoin_1year"
         android:layout_width="match_parent"
@@ -15,7 +64,7 @@
         android:layout_marginTop="20dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toBottomOf="@+id/tv_percentage" />
 
     <TextView
         android:id="@+id/text_view_dados_1year"
@@ -112,4 +161,12 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toBottomOf="@id/text_view_minimo_1year" />
 
+
+        </androidx.constraintlayout.widget.ConstraintLayout>
+
+    </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
+
+
+
 </androidx.constraintlayout.widget.ConstraintLayout>
+
